+++ kaitai/stream.go
@@ -3,15 +3,10 @@ package kaitai
 import (
 	"bufio"
 	"bytes"
-	"compress/zlib"
 	"encoding/binary"
 	"io"
 	"io/ioutil"
 	"math"
-	"math/bits"
-
-	"golang.org/x/text/encoding"
-	"golang.org/x/text/transform"
 )
 
 const APIVersion = 0x0001
@@ -290,43 +285,4 @@ func (k *Stream) ReadBitsInt(n uint8) (val uint64, err error) {
 // FIXME what does this method do?
 func (k *Stream) ReadBitsArray(n uint) error {
 	return nil
-}
-
-func ProcessXOR(data []byte, key []byte) {
-	for i := range data {
-		data[i] ^= key[i%len(key)]
-	}
-}
-
-func ProcessRotateLeft(data []byte, amount int) {
-	for i := range data {
-		data[i] = byte(bits.RotateLeft8(uint8(data[i]), amount))
-	}
-}
-
-func ProcessRotateRight(data []byte, amount int) {
-	ProcessRotateLeft(data, -amount)
-}
-
-func ProcessZlib(in []byte) (out []byte, err error) {
-	b := bytes.NewReader(in)
-
-	// FIXME zlib.NewReader allocates a bunch of memory.  In the future
-	// we could reuse it by using a sync.Pool if this is called in a tight loop.
-	r, err := zlib.NewReader(b)
-	if err != nil {
-		return nil, err
-	}
-
-	return ioutil.ReadAll(r)
-}
-
-func BytesToStr(in []byte, decoder *encoding.Decoder) (out string, err error) {
-	i := bytes.NewReader(in)
-	o := transform.NewReader(i, decoder)
-	d, e := ioutil.ReadAll(o)
-	if e != nil {
-		return "", e
-	}
-	return string(d), nil
 }
